Added a countdown timer to the screen when the level timer ends. Fixed a bug where clicking the Done button from the Store menu resulted in -2 billion minutes remaining.
git-svn-id: https://wiki.csc.calpoly.edu/svn/AsteroidBlaster@701 342171fe-4ba3-422f-b357-09b6f24955fe

+++ Utility/GlobalUtility.cpp
@@ -203,7 +203,7 @@ int minutesRemaining(double secondsRemaining) {
  * left, after the minutes remaining has been calculated.
  */
 int secondsRemainder(double secondsRemaining) {
-   return (int)((int)secondsRemaining)%60;
+   return ((int)secondsRemaining)%60;
 }
 
 void drawCylinder(double radius, double length) {
